requirements/requirements_test.txt
# Test framework
pytest==3.1.0
pytest-django==3.1.2
pytest-cov==2.5.1
pytest-socket==0.2.0
pbr==4.0.2
py==1.5.3
python-dateutil==2.7.3

# Coverage uploads
coverage==4.4.1
coveralls
docopt==0.6.2
requests==2.9.1
responses==0.6.0
cookies==2.2.1

# Mocking, faking, factories and other test helpers
mock==1.3.0
fake-factory==0.6.0
factory-boy==2.6.1
django-factory-boy==1.0.0

# Flake8 checks for PEP-8 formatting and code quality
flake8==3.3.0
enum34==1.1.6
configparser==3.5.0
pyflakes==1.5.0
pep8==1.7.0
pycodestyle==2.3.1
# Checks for cyclomatic complexity
mccabe==0.6.1
# Checks for naming conventions
pep8-naming==0.4.1
# Checks for the use of print statements
flake8-print==2.0.2
# Checks for the use of debug statements (pdb, ...)
flake8-debugger==1.4.0
# Checks for things that should be comprehensions
flake8-comprehensions==1.2.1
# Checks for non-existent mock methods
flake8-mock==0.3
# Checks for mutable function arguments
flake8-mutable==1.1.0
# Checks for TODO statements (and similar)
flake8-todo==0.7
pycodestyle==2.3.1
# Checks for uses of double quotes (without containing single quotes)
flake8-quotes==0.9.0
# Checks for uses of old-style string formatting (using % instead of .format())
flake8-pep3101==1.0
# Checks for the overriding of builtins
flake8-builtins-unleashed==1.2
# Checks for uses of Django-style asserts
flake8-pytest==1.3
# Checks for module-level imports
flake8-module-imports==1.1
# Checks for usages of django.translation.activate()
flake8-translation-activate==1.0.2
# Checks for imports with an alias of _
flake8-ugettext-alias==1.1
# Checks for imports of the User model (instead of get_user_model() or settings.AUTH_USER_MODEL)
flake8-user-model==1.1
